+++ pytgbot/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import json
 import logging
 from .bot import Bot